api/callcentersite/requirements/test.txt
# callcentersite/requirements/test.txt
# ============================================
# FRAMEWORK DE TESTING
# Instaladas en onCreateCommand (una sola vez)
# ============================================

# Instalar primero las dependencias base
-r base.txt

# ============================================
# CORE TESTING
# ============================================
pytest>=8.0.0
pytest-django>=4.7.0

# ============================================
# COVERAGE
# ============================================
pytest-cov>=4.1.0
coverage>=7.3.2

# ============================================
# TEST EXECUTION
# ============================================
# Parallel execution
pytest-xdist>=3.5.0

# Watch mode for TDD
pytest-watch>=4.2.0

# ============================================
# TEST DATA & MOCKING
# ============================================
# Factories
factory-boy>=3.3.0
faker>=20.1.0

# Mocking
pytest-mock>=3.12.0
responses>=0.24.1

# Time manipulation
freezegun>=1.4.0

# ============================================
# TEST UTILITIES
# ============================================
# Better assertions
pytest-clarity>=1.0.1

# Performance profiling
pytest-benchmark>=4.0.0

# Better output
pytest-sugar>=0.9.7
pytest-html>=4.1.1

# ============================================
# ENVIRONMENT
# ============================================
python-dotenv>=1.0.0

mysql-connector-python
psycopg2-binary
psutil